Some minor refactoring #mod-6
