+++ modulus.api/StandardMod11.cs
@@ -3,26 +3,16 @@ using Modulus.Shared.Models;
 
 namespace Modulus.api
 {
-    public class StandardMod11 : ModulusCalcBase
+    public class StandardMod11 : Standard
     {
         public StandardMod11(AccountInfo accountInfo, WeightItem weightItem) : base(accountInfo, weightItem)
         {
+            ModulusNumber = 11;
         }
 
         public override int Calculate()
         {
-            var num = 0;
-            for (var i = 0; i < 6; i++)
-            {
-                num += int.Parse(this.AccountInfo.SortCode.Substring(i, 1)) * WeightItem.Notation[i];
-            }
-                
-            for (var i = 0; i < 8; i++)
-            {
-                num += int.Parse(this.AccountInfo.AccountNumber.Substring(i, 1)) * WeightItem.Notation[6+i];
-            }
-            
-            var remainder = num % 11;
+            int remainder = base.Calculate();
 
             // Exception 4, Modulus 11 only
